@@ -1,29 +1,40 @@
 {
   "name": "filesystem",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "description": "FileSystem",
   "main": "main.js",
   "scripts": {
     "start": "electron main.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "electron-packager . FileSystem --platform=win32 --arch=all --version=0.32.3"
+    "build": "electron-packager . FileSystem --platform=win32 --asar --arch=ia32"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Ruiming/FileSystem.git"
+  },
+  "bugs": {
+    "url": "https://github.com/Ruiming/FileSystem/issues"
+  },
+  "homepage": "https://github.com/Ruiming/FileSystem#readme",
   "author": "Ruiming Zhuang",
   "license": "ISC",
-  "devDependencies": {
+  "dependencies": {
     "base64-img": "^1.0.3",
-    "bower": "^1.7.9",
-    "chokidar": "^1.6.0",
-    "electron": "^0.4.1",
-    "electron-packager": "^7.0.3",
-    "electron-prebuilt": "^1.2.1",
-    "electron-rebuild": "^1.1.5",
-    "ffi": "^2.0.0",
-    "gulp": "^3.9.1",
-    "gulp-plumber": "^1.1.0",
-    "gulp-sass": "^2.2.0",
     "iconv-lite": "^0.4.13",
-    "node-wmic": "^1.0.3",
-    "watch": "^0.19.1"
+    "node-wmic": "^1.0.6",
+    "watch": "^0.19.1",
+    "electron": "^1.3.2"
+  },
+  "devDependencies": {
+    "bower": "^1.7.9",
+    "electron-packager": "^7.5.1",
+    "electron-prebuilt": "^1.3.2",
+    "gulp": "^3.9.0",
+    "gulp-babel": "^6.1.2",
+    "gulp-concat": "^2.6.0",
+    "gulp-plumber": "^1.1.0",
+    "gulp-sass": "^2.3.2",
+    "babel-core": "^6.13.2",
+    "babel-preset-es2015": "^6.13.2"
   }
 }
