@@ -1,23 +1,34 @@
-var routeApp = angular.module('asd',['ui.router','ui.bootstrap', 'angularBootstrapMaterial', 'ngAnimate']);
+(function() {
+    'use strict';
 
-routeApp.config(['$stateProvider', '$urlRouterProvider', function($stateProvider, $urlRouterProvider){
+    angular
+        .module('app', [
+            'ui.router',
+            'ui.bootstrap',
+            'angularBootstrapMaterial',
+            'ngAnimate'
+        ])
+        .config(config);
 
-    $urlRouterProvider.otherwise("/");
-    $stateProvider
-        .state('index', {
-            url: '/',
-            views: {
-                'system': {
-                    templateUrl: 'app/templates/index.html',
-                    controller: 'indexCtrl',
-                    controllerUrl: 'app/controllers/indexCtrl.js'
-                },
-                'files': {
-                    templateUrl: 'app/templates/file.html',
-                    controller: 'fileCtrl',
-                    controllerUrl: 'app/controllers/fileCtrl.js'
+    config.$inject = ['$stateProvider', '$urlRouterProvider'];
+
+    function config($stateProvider, $urlRouterProvider) {
+        $urlRouterProvider.otherwise("/");
+        $stateProvider
+            .state('index', {
+                url: '/',
+                views: {
+                    'system': {
+                        templateUrl: 'app/templates/index.html',
+                        controller: 'IndexCtrl',
+                        controllerUrl: 'app/controllers/IndexCtrl.js'
+                    },
+                    'files': {
+                        templateUrl: 'app/templates/file.html',
+                        controller: 'FileCtrl',
+                        controllerUrl: 'app/controllers/fileCtrl.js'
+                    }
                 }
-            }
-        });
-
-}]);
+            });
+    }
+}());
